Add basic csrf protection on all user pages

+++ src/Controller/Media/GalleryController.php
@@ -61,9 +61,11 @@ class GalleryController extends AbstractController
     /**
      * @Route("/media/gallery/comment", methods="POST")
      */
-    public function getComment(Request $request){
+    public function comment(Request $request){
         $this->denyAccessUnlessGranted(Permission::IS_LOGIN);
         $this->denyAccessUnlessGranted(Permission::IS_AUTHENTICATED);
+        if(!$this->verfityCsrfToken($request->request->get("_csrf"),AbstractController::CSRF_MEDIA_GALLERY))
+            return $this->response->response("csrf.invalid",Response::HTTP_BAD_REQUEST);
         $content = $request->request->get("content");
         $repo = $this->getDoctrine()->getManager()->getRepository(Gallery::class);
         $comment = new Comment();
@@ -81,6 +83,8 @@ class GalleryController extends AbstractController
      */
     public function like(Request $request){
         $this->denyAccessUnlessGranted(Permission::IS_LOGIN);
+        if(!$this->verfityCsrfToken($request->request->get("_csrf"),AbstractController::CSRF_MEDIA_GALLERY))
+            return $this->response->response("csrf.invalid",Response::HTTP_BAD_REQUEST);
         $repo = $this->getDoctrine()->getManager()->getRepository( Gallery::class);
         $gallery = $repo->getGallery($request->request->get("id"));
         $gallery->like($this->getUser());
@@ -145,6 +149,8 @@ class GalleryController extends AbstractController
         $em = $this->getDoctrine()->getManager();
         $repo = $em->getRepository(Comment::class);
         if($request->request->has("delete")){
+            if(!$this->verfityCsrfToken($request->request->get("_csrf"),AbstractController::CSRF_MEDIA_GALLERY))
+                return $this->response->response("csrf.invalid",Response::HTTP_BAD_REQUEST);
             $items = json_decode($request->request->get("delete"),true);
             foreach ($items as $item){
                 $em->remove($repo->findOneBy(["id"=>$item]));
